A basic notes keeping app
notes.js contain the actions in the notes-app
app.js creates the functions required to perform actions on notes-app.
notes.json is where the notes are stored, retrieved from, deleted added, and removed.
Will act a react based frontend soon.

IMPORTANT: 
This was developed as a part of Andrew Mead's The Complete Node.js Developer Course(3rd edition on Udemy). he is one of the finest backend tutors in the MERN stack.